refactor: extract components and hooks
