+++ src/lib/utils.ts
@@ -1,0 +1,3 @@
+export function clsx(...parts: Array<string | false | null | undefined>) {
+  return parts.filter(Boolean).join(" ");
+}
